parent access login files
home_closed.html is edited to provide a message to parents when the portal is closed
for grading or maintenance. the file is then renamed to home.html, and appears as the login screen,
but without a place for entering the username and password.

logonhelp.html is the link parents get when they ask for help.